complete
It is basically a methodology like other programming methodologies e.g structure programming, modular programming, procedural programming or aspect oriented programming etc but it is one of the most famous and well known paradigm of programing language. All the latest languages supports OOP. It is more related to internal designing of a software. It works with the programmer of a software